monitor-ssl-k8s/requirements.txt
pymongo==4.6.3
PyYAML==6.0.1
requests==2.31.0
pyTelegramBotAPI==4.16.1
schedule==1.2.1